% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{plant.arr}
\alias{initiate.plants}
\alias{plant.arr}
\alias{plant.timestep}
\alias{run.eco}
\alias{survive}
\title{Inserts plant species into the terrain matrix.}
\usage{
initiate.plants(repduc, survive, names = NULL)

survive(cell, info)

plant.timestep(plants, info)

run.eco(terra, timesteps, repduc, survive)
}
\arguments{
\item{repduc}{the probability that the corresponding plant species
reproduces.}

\item{survive}{the probability that the corresponding plant species
survives.}

\item{names}{argument that links a logical value to its corresponding
numerical value for survive and reproduction.}

\item{cell}{argument used to determine the given content of a matrix
cell.}

\item{info}{A \code{\link{run.eco}} object}

\item{plants}{argument to call the 'plants' array in order to be
manipulated.}

\item{terra}{terrain matrix generated by terrain.r}

\item{timesteps}{amount of 'timesteps' being applied. Gives the
third dimension of the plants array.}
}
\value{
initiate.plants objects

the alterations done by plant.timestep

plants array with simulated species inserted.
}
\description{
Simulation of plant species placed on the terrain matrix
   built in 'terrain.r'. Where there is terrain, the plant species
   initially survive, then reproduce, and eventually compete with each
   other.
}
\examples{
   plants<-run.eco(terra=terrain.wrapper(6,TRUE),10,repduc=c(.4, .6),
   survive=c(.58,.65))
}

