% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terrain}
\alias{diamond.square.step}
\alias{diamond.step}
\alias{square.step}
\alias{terrain}
\alias{terrain.wrapper}
\alias{world.dim}
\title{Topographical terrain represented by numeric values in each cell of
a matrix.}
\usage{
world.dim(dim)

diamond.step(terra)

square.step(terra)

diamond.square.step(terra)

terrain.wrapper(dim, lakes)
}
\arguments{
\item{dim}{Size of the grid will be 2^dim+1 (e.g. dim=3; 9X9 grid).}

\item{terra}{argument of the matrix 'terra' that is manipulated by 
the corresponding functions.}

\item{lakes}{logical that determines whether or not the negative values
are made to be NAs aka underwater (default: TRUE).}
}
\value{
a terrain matrix; numeric elements indicate heights, and
   NAs indicate cells filled with water.
}
\description{
Creates a matrix that is filled beginning with the corners by
   a diamond-square step algorithim. The values in each cell represent
   terrain altitudes. Lakes are signified by NAs within the matrix if
   the user chooses to implement them in terrain.
}
\examples{
   terra<-terrain.wrapper(6, TRUE)
   image(terra)
}

